feat: new build

+++ src/main/java/edu/sjsu/mseto/p1_gui/controller/BankingSystemRestController.java
@@ -142,8 +142,8 @@ public class BankingSystemRestController {
     @GetMapping("/api/addInterest")
     public String addInterest(@RequestParam(value = "savingsRate") String savingsRate, @RequestParam(value = "checkingRate") String checkingRate) {
         try {
-            if (Integer.parseInt(savingsRate) <= 0) throw new Exception("Invalid savings rate");
-            if (Integer.parseInt(checkingRate) <= 0) throw new Exception("Invalid checking rate");
+            if (Float.parseFloat(savingsRate) <= 0) throw new Exception("Invalid savings rate");
+            if (Float.parseFloat(checkingRate) <= 0) throw new Exception("Invalid checking rate");
             return BankingSystem.addInterest(savingsRate, checkingRate);
         } catch (Exception e) {
             e.printStackTrace();
